Rename `kit` to `boot`
