+++ src/boot/definition/component-registry.impl.ts
