+++ src/boot/globals/bootstrap-window.ts
