+++ src/boot/globals/element-adapter.ts
