+++ src/boot/globals/index.ts
