+++ src/component/definition/element-def.ts
@@ -3,8 +3,8 @@
  */
 import { SingleContextKey, SingleContextRef } from 'context-values';
 import { QualifiedName } from 'namespace-aliaser';
+import { BootstrapWindow } from '../../boot/globals';
 import { Class } from '../../common';
-import { BootstrapWindow } from '../../kit/globals';
 import { ComponentDef } from '../component-def';
 import { DefinitionContext__key } from './definition.context.key.impl';
