+++ src/feature/feature-def.ts
@@ -2,10 +2,10 @@
  * @module @wesib/wesib
  */
 import { ContextValueSpec } from 'context-values';
+import { BootstrapContext } from '../boot';
 import { ArraySet, Class, mergeFunctions, MetaAccessor } from '../common';
 import { ComponentContext } from '../component';
 import { DefinitionContext } from '../component/definition';
-import { BootstrapContext } from '../kit';
 import { FeatureContext } from './feature-context';
 
 /**
