+++ src/index.ts
@@ -1,6 +1,9 @@
 /**
  * @module @wesib/wesib
  */
+export * from './boot';
+export * from './boot/bootstrap';
+export * from './boot/globals';
 export * from './common';
 export * from './component';
 export * from './component/definition';
@@ -11,6 +14,3 @@ export * from './feature/dom-properties';
 export * from './feature/render';
 export * from './feature/shadow-dom';
 export * from './feature/state';
-export * from './kit';
-export * from './kit/bootstrap';
-export * from './kit/globals';
